# docker_images/csharp/wrapper/src/Controllers/ControlApi.cs
/*
 * Azure IOT End-to-End Test Wrapper Rest Api
 *
 * REST API definition for End-to-end testing of the Azure IoT SDKs.  All SDK APIs that are tested by our E2E tests need to be defined in this file.  This file takes some liberties with the API definitions.  In particular, response schemas are undefined, and error responses are also undefined.
 *
 * OpenAPI spec version: 1.0.0
 *
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{
    /// <summary>
    ///
    /// </summary>
    public class ControlApiController : Controller
    {
        // Added 1 line in merge
        internal ControlGlue control_glue = new ControlGlue();

        /// <summary>
        /// verify that the clients have cleaned themselves up completely
        /// </summary>

        /// <response code="200">OK</response>
        [HttpPut]
        [Route("/control/cleanup")]
        [ValidateModelState]
        [SwaggerOperation("ControlCleanup")]
        public virtual IActionResult ControlCleanup()
        {
            // Changed impl in merge
            control_glue.CleanupResourcesAsync().Wait();
            return StatusCode(200);

        }

        /// <summary>
        /// Get capabilities for the objects in this server
        /// </summary>

        /// <response code="200">OK</response>
        [HttpGet]
        [Route("/control/capabilities")]
        [ValidateModelState]
        [SwaggerOperation("ControlGetCapbilities")]
        [SwaggerResponse(statusCode: 200, type: typeof(Object), description: "OK")]
        public virtual IActionResult ControlGetCapabilities()
        {
            // Changed impl in merge
            Task<object> t = control_glue.GetCapabilities();
            t.Wait();
            return new ObjectResult(t.Result);
        }

        /// <summary>
        /// log a message to output
        /// </summary>

        /// <param name="logMessage"></param>
        /// <response code="200">OK</response>
        [HttpPut]
        [Route("/control/message")]
        [ValidateModelState]
        [SwaggerOperation("ControlLogMessage")]
        public virtual IActionResult ControlLogMessage([FromBody]LogMessage logMessage)
        {
            // Changed impl in merge
            control_glue.PrintMessageAsync(logMessage.Message).Wait();
            return StatusCode(200);
        }

        /// <summary>
        /// send an arbitrary command
        /// </summary>

        /// <param name="cmd">command string</param>
        /// <response code="200">OK</response>
        [HttpPut]
        [Route("/control/command")]
        [ValidateModelState]
        [SwaggerOperation("ControlSendCommand")]
        public virtual IActionResult ControlSendCommand([FromQuery][Required()]string cmd)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);


            throw new NotImplementedException();
        }

        /// <summary>
        /// set flags for the objects in this server to use
        /// </summary>

        /// <param name="flags"></param>
        /// <response code="200">OK</response>
        [HttpPut]
        [Route("/control/flags")]
        [ValidateModelState]
        [SwaggerOperation("ControlSetFlags")]
        public virtual IActionResult ControlSetFlags([FromBody]Object flags)
        {
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);


            throw new NotImplementedException();
        }
    }
}