# docker_images/csharp/wrapper/src/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API definition for End-to-end testing of the Azure IoT SDKs.  All SDK APIs that are tested by our E2E tests need to be defined in this file.  This file takes some liberties with the API definitions.  In particular, response schemas are undefined, and error responses are also undefined.",
    "version" : "1.0.0",
    "title" : "Azure IOT End-to-End Test Wrapper Rest Api"
  },
...
  "schemes" : [ "https" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/control/message" : {
      "put" : {
        "tags" : [ "control" ],
        "summary" : "log a message to output",
        "operationId" : "Control_LogMessage",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "logMessage",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/logMessage"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/control/flags" : {
      "put" : {
        "tags" : [ "control" ],
        "summary" : "set flags for the objects in this server to use",
        "operationId" : "Control_SetFlags",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "flags",
          "required" : true,
          "schema" : {
            "type" : "object"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/control/command" : {
      "put" : {
        "tags" : [ "control" ],
        "summary" : "send an arbitrary command",
        "operationId" : "Control_SendCommand",
        "parameters" : [ {
          "name" : "cmd",
          "in" : "query",
          "description" : "command string",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "logMessage" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "text of message"
        }
      },
      "description" : "message from the test script to output to the log"
    },
...
  }
}